+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## [Unreleased]
 
+## [0.2.2] - 2024-01-26
+### Added
+- Support for having conditions
+- Literal condition values for first party mode
+
+### Changed
+- Order name can be unescaped in first party mode
+
 ## [0.2.1] - 2023-07-03
 ### Added
 - Null and NotNull where comparators
@@ -24,7 +32,8 @@
 - Nette database schema loader
 - Nette database resolvers
 
-[Unreleased]: https://github.com/efabrica-team/nette-graphql/compare/0.2.1...main
+[Unreleased]: https://github.com/efabrica-team/nette-graphql/compare/0.2.2...main
+[0.2.2]: https://github.com/efabrica-team/nette-graphql/compare/0.2.1...0.2.2
 [0.2.1]: https://github.com/efabrica-team/nette-graphql/compare/0.2.0...0.2.1
 [0.2.0]: https://github.com/efabrica-team/nette-graphql/compare/0.1.0...0.2.0
 [0.1.0]: https://github.com/efabrica-team/nette-graphql/compare/0.0.0...0.1.0